Adds Elasticsearch 5.x support via storage type: elasticsearch-http
This allows use of Elasticsearch 5.x and its notable ingest pipeline
feature when using `zipkin-storage-elasticsearch-http`.

It is important to note that `zipkin-storage-elasticsearch` remains
pinned to ES 2.x libraries as they are compile incompatible with 5.x.
In other words, you must use http if you want to use ES 5 (for now).

Version detection is implemented in order to choose the coirrect index
template format for the major version number. The implementation of
such is string manip, as it was less work than making a new type.

This adds a new parameter `ES_PIPELINE` which allows you to manipulate
the json sent by Zipkin collector before it is indexed. This could be
used for many things including cleaning service names or adding ingest
timestamps.

Integration tests run version 2.x on CircleCI and 5.x on Travis